Finished making /tag subcommands
/tag now has every subcommand that .tag has. Also includes some changes to error handling to work with slash commands.

+++ cogs/tags.py
@@ -184,11 +184,13 @@ class Choices(str, Enum):
     Edit = 'edit'
     Delete = 'delete'
     List = 'list'
+    Make = 'make'
     Raw = 'raw'
     Rename = 'rename'
     Alias = 'alias'
     Info = 'info'
     Transfer = 'transfer'
+    Tags = 'tags'
 
 class Tags(AceMixin, commands.Cog):
     """Store and bring up text using tags. Tags are unique to each server."""
@@ -319,6 +321,11 @@ class Tags(AceMixin, commands.Cog):
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to delete.")
                 _, record = await TagEditConverter(allow_mod=True).convert(inter, query.split(ZWS)[0])
+                inter.perms = inter.channel.permissions_for(inter.guild.me)
+                if not await AceContext.prompt(inter,
+                    title="Are you sure?", prompt="This will delete the tag permanently."
+                ):
+                    raise commands.CommandError("Tag deletion aborted.")
                 await self.db.execute("DELETE FROM tag WHERE id=$1", record.get("id"))
                 await inter.send(f"Tag '{record.get('name')}' deleted.")
                 pass
@@ -345,34 +352,182 @@ class Tags(AceMixin, commands.Cog):
                 p.member = member
 
                 await p.go()
+            case Choices.Make:
+                def msg_check(message):
+                    return message.channel is inter.channel and message.author is inter.author
+                name = None
+                content = None
+
+                self.set_tag_being_made(inter, name)
+
+                name_prompt = "What would you like the name of your tag to be?"
+
+                inter.message = await inter.send("Hi there! " + name_prompt)
+
+                while True:
+                    try:
+                        message = await inter.bot.wait_for("message", check=msg_check, timeout=360.0)
+                    except asyncio.TimeoutError:
+                        await inter.send(
+                            "The tag make command timed out. Please try again by doing "
+                            "`/tag subcom:make`"
+                        )
+
+                        self.unset_tag_being_made(inter)
+                        return
+
+                    if message.content == "/abort":
+                        await inter.send("Tag creation aborted.")
+                        self.unset_tag_being_made(inter)
+                        return
+
+                    if name is None:
+                        try:
+                            old_message = inter.message
+                            inter.message = message
+                            await tag_create_converter.convert(inter, message.content)
+                            inter.message = old_message
+                        except commands.CommandError as exc:
+                            await inter.send("Sorry! {} {}".format(str(exc), name_prompt))
+                            continue
+
+                        name = message.content.lower()
+                        self.set_tag_being_made(inter, name)
+
+                        await inter.send(
+                            "Great! The tag name is `{}`. What would you like the tags content to be?\n".format(
+                                name
+                            )
+                            + "You can abort the tag creation by sending `/abort` at any time."
+                        )
+
+                        continue
+
+                    if content is None:
+                        old_message = inter.message
+                        inter.message = message
+                        content = await self.craft_tag_contents(inter, message.content)
+                        inter.message = old_message
+                        break
+
+                self.unset_tag_being_made(inter)
+                await self.create_tag(inter, name, content)
+
+                await inter.send(
+                    "Tag `{0}` created! Bring up the tag contents by doing `/tag {0}`".format(
+                        name
+                    )
+                )
             case Choices.Raw:
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to see the contents of.")
                 _, record = await TagViewConverter().convert(inter, query.split(ZWS)[0])
-                pass
+                await inter.send(
+                    disnake.utils.escape_markdown(record.get("content")),
+                    allowed_mentions=disnake.AllowedMentions.none(),
+                )
             case Choices.Rename:
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to rename.")
                 _, record = await TagEditConverter(allow_mod=True).convert(inter, query.split(ZWS)[0])
                 if string is None:
                     raise commands.CommandError("Please input the new tag name in the string paramater.")
+                await self.db.execute("UPDATE tag SET name=$2 WHERE id=$1", record.get("id"), string)
+
+                await inter.send(f"Tag '{record.get('name')}' renamed to '{string}'.")
             case Choices.Alias:
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to create an alias for.")
                 _, record = await TagEditConverter(allow_mod=True).convert(inter, query.split(ZWS)[0])
                 if string is None:
                     raise commands.CommandError("Please input the new tag alias in the string paramater.")
+                await self.db.execute("UPDATE tag SET alias=$2 WHERE id=$1", record.get("id"), string)
+
+                if string is None:
+                    await inter.send(f"Alias cleared for '{record.get('name')}'")
+                else:
+                    await inter.send(f"Alias for '{record.get('name')}' set to '{string}'")
             case Choices.Info:
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to see the statistics for.")
                 _, record = await TagViewConverter().convert(inter, query.split(ZWS)[0])
-                pass
+
+                owner = inter.guild.get_member(record.get("user_id"))
+                if owner is None:
+                    nick = "Unknown User"
+                    avatar = inter.guild.icon or None
+                else:
+                    nick = owner.display_name
+                    avatar = owner.display_avatar.url
+
+                e = disnake.Embed(
+                    description=f"**{record.get('name')}**",
+                )
+
+                e.set_author(name=nick, icon_url=avatar)
+                e.add_field(name="Owner", value=owner.mention if owner else nick)
+
+                rank = await self.db.fetchval(
+                    "SELECT COUNT(id) FROM tag WHERE guild_id=$1 AND uses > $2",
+                    inter.guild.id,
+                    record.get("uses") + 1,
+                )
+
+                e.add_field(name="Rank", value=f"#{rank + 1}")
+
+                e.add_field(name="Uses", value=record.get("uses"))
+
+                alias = record.get("alias")
+                created_at = record.get("created_at")
+                viewed_at = record.get("viewed_at")
+                edited_at = record.get("edited_at")
+
+                if alias is not None:
+                    e.add_field(name="Alias", value=alias)
+
+                e.add_field(name="Created at", value=pretty_datetime(created_at))
+
+                if viewed_at:
+                    e.add_field(name="Last viewed at", value=pretty_datetime(viewed_at))
+
+                if edited_at:
+                    e.add_field(name="Last edited at", value=pretty_datetime(edited_at))
+
+                await inter.send(embed=e)
             case Choices.Transfer:
                 if query is None:
                     raise commands.CommandError("Please input the tag you want to transfer.")
                 _, record = await TagEditConverter(allow_mod=True).convert(inter, query.split(ZWS)[0])
                 if member is None:
                     raise commands.CommandError("Please input the new tag owner in the member paramater.")
+                if member.bot:
+                    raise commands.CommandError("Can't transfer tag to bot.")
+
+                if record.get("user_id") == member.id:
+                    raise commands.CommandError("User already owns tag.")
+
+                inter.perms = inter.channel.permissions_for(inter.guild.me)
+                prompt = AceContext.prompt(inter,
+                    title="Tag transfer request",
+                    prompt=f"{inter.author.mention} wants to transfer ownership of the tag '{query}' to you.\n\nDo you accept?",
+                    user_override=member,
+                )
+
+                if not await prompt:
+                    raise commands.CommandError("Tag transfer aborted.")
+
+                res = await self.db.execute(
+                    "UPDATE tag SET user_id=$1 WHERE id=$2", member.id, record.get("id")
+                )
+
+                if res == "UPDATE 1":
+                    await inter.send(
+                        "Tag '{}' transferred to '{}'".format(record.get("name"), member.display_name)
+                    )
+                else:
+                    raise commands.CommandError("Unknown error occured.")
+            case Choices.Tags:
+                await inter.application_command.callback(self, inter, subcom=Choices.List,member=member or inter.author)
